refactor medication info
To make it more "reasonaboutable" and theoretically more testable
