+++ src/views/MedicationInfo/index.js
@@ -2,19 +2,13 @@ import React, { Component } from 'react';
 import styled from 'styled-components';
 import uuid from 'uuid';
 
+import MedSection from './MedSection';
+import MedTitle from './MedTitle';
 import data from './../../databaseMeds';
 import { PageContainer } from './../../styled';
-import astronaut from './assets/astronaut.svg';
-import calendar from './assets/calendar.svg';
-import robot from './assets/robot.svg';
-import telescope from './assets/telescope.svg';
 
-const icons = {
-  name: astronaut,
-  pronunciation: telescope,
-  use: calendar,
-  'side-effects': robot,
-};
+const name = 'name';
+const pronunciation = 'pronunciation';
 
 const MedicationInfoWrapper = styled.div`
   padding-left: 1rem;
@@ -23,53 +17,7 @@ const MedicationInfoWrapper = styled.div`
   background-color: #052136;
 `;
 
-const PageTitle = styled.h1`
-  ${/* color: #052136; */ ''}
-  color: #fff;
-  font-weight: 700;
-  text-align: center;
-  text-transform: uppercase;
-`;
-
-const Pronunciation = styled.p`
-  color: #fff;
-  font-weight: 700;
-  text-align: center;
-`;
-
-const InfoList = styled.ul`
-`;
-
-const InfoItem = styled.li`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 2rem;
-  padding: 1rem;
-  border-radius: 3px;
-  ${/* background-color: #fafafa; */ ''}
-`;
-
-const SectionIcon = styled.img`
-  margin-bottom: 1rem;
-  width: 7rem;
-  height: auto;
-`;
-
-const SectionTitle = styled.h2`
-  margin-bottom: 1rem;
-  text-align: center;
-  color: #fff;
-  font-weight: 700;
-`;
-
-const Description = styled.div`
-  color: #f1f1f1;
-  padding: 0.5rem;
-  > div {
-    text-align: center;
-  }
-`;
+const InfoList = styled.ul``;
 
 const Effect = styled.li`
   list-style-type: circle;
@@ -77,61 +25,47 @@ const Effect = styled.li`
 `;
 
 class MedicationInfo extends Component {
-  capitalizeFirst = word => word.charAt(0).toUpperCase() + word.slice(1);
+  state = {
+    medication: '',
+    drug: {},
+  };
+  componentDidMount() {
+    const { medication } = this.props.match.params;
+    const drug = data.find(drug => drug.name === medication);
+    this.setState({
+      medication,
+      drug,
+    });
+  }
+
   renderSideEffects = sideEffects => (
     <ul>
       {sideEffects.map(effect => <Effect key={uuid()}>{effect}</Effect>)}
     </ul>
   );
 
-  renderPronunciation = drugs => (
-    <div>
-      <p>{drugs.name}</p><p>{drugs.pronunciation}</p>
-    </div>
-  );
-
   render() {
-    const medication = this.props.match.params.medication;
+    const { drug } = this.state;
     return (
       <PageContainer>
         <MedicationInfoWrapper>
           <InfoList>
-            {data.map(drug => {
-              if (drug.name === medication) {
-                return Object.keys(drug).map(detail => {
-                  if (detail === 'name') {
-                    return (
-                      <InfoItem>
-                        <PageTitle>
-                          {drug.name}
-                        </PageTitle>
-                        <Pronunciation>
-                          {drug.pronunciation}
-                        </Pronunciation>
-                      </InfoItem>
-                    );
-                  }
-                  if (detail === 'pronunciation') {
-                    return;
-                  }
+            {Object.keys(drug).map(detail => {
+              switch (detail) {
+                case name:
+                  return <MedTitle key={uuid()} drug={drug} />;
+                case pronunciation:
+                  return null;
+                default:
                   return (
-                    <InfoItem key={uuid()}>
-                      <SectionIcon src={icons[detail]} />
-                      <SectionTitle>
-                        {this.capitalizeFirst(detail)}
-                      </SectionTitle>
-                      <Description>
-                        {detail === 'side-effects'
-                          ? this.renderSideEffects(drug[detail])
-                          : detail === 'name'
-                              ? this.renderPronunciation(drug)
-                              : drug[detail]}
-                      </Description>
-                    </InfoItem>
+                    <MedSection
+                      key={uuid()}
+                      drug={drug}
+                      detail={detail}
+                      renderSideEffects={this.renderSideEffects}
+                    />
                   );
-                });
               }
-              return null;
             })}
           </InfoList>
         </MedicationInfoWrapper>
